requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
numpy
openpyxl
huggingface_hub
